pyproject.toml
[tool.poetry]
name = "questionnaire"
version = "0.1.0"
description = ""
authors = ["Slmpnv <slmpnvmail@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytelegrambotapi = "^4.14.1"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
